% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSExtract.R
\name{ridge.swale}
\alias{ridge.swale}
\title{Ridge-Swale morphometry extraction using surface normal vectors}
\usage{
ridge.swale(inEl, inX, inY)
}
\arguments{
\item{inEl}{Avg. Elevation post unit vector extraction. (SpatRaster)}

\item{inX}{X-component of the surface normal vectors. (SpatRaster)}

\item{inY}{Y-component of the surface normal vectors. (SpatRaster)}
}
\value{
All combinations of ridge swale patterns, returns a list containing both
matrix and raster data structures.
}
\description{
Ridge-Swale morphometry extraction using surface normal vectors
}
