% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{single.RM}
\alias{single.RM}
\title{Remove single cell feature noise}
\usage{
single.RM(rs.List)
}
\arguments{
\item{rs.List}{List of ridge and swale feature matrices and rasters (from ridge.swale function)}
}
\value{
All combinations of ridge swale patterns with 'noise' removed, returns a list containing both
matrix and raster data structures.
}
\description{
Remove single cell feature noise
}
