% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{UV.RS.extract}
\alias{UV.RS.extract}
\title{Single function to extract unit vectors and cleaned ridge-swale features}
\usage{
UV.RS.extract(inRas, win = 5, rm.S = TRUE)
}
\arguments{
\item{inRas}{Input DEM. (SpatRast)}

\item{win}{Window size, should be an odd value greater than 1.(default 5)}

\item{rm.S}{Boolean choice if you want to remove 'noise' from RS features. (default = TRUE)}
}
\value{
List of unit vector and ridge swale feature lists.
}
\description{
Single function to extract unit vectors and cleaned ridge-swale features
}
