% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNVectors.R
\name{unitVectors}
\alias{unitVectors}
\title{Surface Normal Unit Vectors of DEM}
\usage{
unitVectors(inRas, win = 5)
}
\arguments{
\item{inRas}{Input DEM SpatRaster.}

\item{win}{Window size, should be an odd value greater than 1.(default 5)}
}
\value{
XYZ components of the unit vectors, avg. elevation,
and approximations of surface area and slope, returns a list containing both
matrix and raster data structures.
}
\description{
Surface Normal Unit Vectors of DEM
}
